Data-Visualization-of-US-Presidential-Results/requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
census==0.8.15
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
citipy==0.0.5
click==7.1.2
colorama==0.4.3
configparser==5.0.1
crayons==0.4.0
cx-Oracle==8.1.0
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
geojson==2.5.0
gmaps==0.9.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
jellyfish==0.6.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kdtree==0.16
kiwisolver==1.2.0
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.18.5
OpenBlender==2.3
openweathermapy==0.6.6
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.2.0
pip==20.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.1
pymongo==3.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
scipy==1.5.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
splinter==0.14.0
SQLAlchemy==1.3.17
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
us==2.0.2
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.0
zipp==3.2.0
zope.interface==5.1.0
